style: 将tti移到`tool_image.ts`内 #1
文件太多不好管理
